refactor: move `PasswordInput` to use `Islet` architecture
As part of moving to the Islands (`Islet`) architecture, we're migrating
all client side code to use this method (see

This commit moves the `PasswordInput` component to this architecture, so
it's loaded only on pages that require it.
