+++ src/client/components/PasswordInput.stories.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { Meta } from '@storybook/react';
 
-import { PasswordInput, PasswordInputProps } from './PasswordInput';
+import { PasswordInput, PasswordInputProps } from './PasswordInput.importable';
 
 export default {
   title: 'Components/PasswordInput',
